+++ ScreenCapture/Util.cpp
@@ -70,7 +70,7 @@ std::string ConvertToUTF8(const std::wstring& wstr)
 
 std::string ConvertToUTF8(const LPWSTR& wstr)
 {
-    const int count = WideCharToMultiByte(CP_UTF8, 0, wstr, wcslen(wstr), NULL, 0, NULL, NULL);
+    const int count = WideCharToMultiByte(CP_UTF8, 0, wstr, (int)wcslen(wstr), NULL, 0, NULL, NULL);
     std::string str(count, 0);
     WideCharToMultiByte(CP_UTF8, 0, wstr, -1, &str[0], count, NULL, NULL);
     return str;
